requirements.txt
pymupdf>=1.27
regex>=2024.0
tqdm>=4.66
openai>=1.50
python-dotenv>=1.0
neo4j>=5.20
streamlit>=1.32
